% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blotPlot.R
\name{blotPlot}
\alias{blotPlot}
\title{Generate violin plots with a high density fit (i.e. blot plots).}
\usage{
blotPlot(cellData, genes, groups = "GroupID", colors, cols = 3, singleGroup,
  singleColor = "red", order = TRUE, save = FALSE)
}
\arguments{
\item{cellData}{ExpressionSet object created with readCells (and preferably
transformed with prepCells).}

\item{genes}{Vector of character strings specifying the names of genes to
plot.  All gene names must be genes that are present in the expression
matrix.}

\item{groups}{Character string specifying the title of the column in pData
that contains the names of the groups to which each sample belongs.  The
blots representing each group in the plots for each gene will be colored
accordingly. The column length should be the same length as the number of
samples.}

\item{colors}{Character string specifying the title of the column in pData
that contains the colors to be used for plotting.  The number of colors
should be equal to the number of levels in groups.  The order of colors
corresponds to the order of factor levels in groups.  If nothing is
specified, random colors from the rainbow function are assigned.}

\item{cols}{Integer specifying the number of columns in which to arrange the
blot plots.  The number of cols should be less than or equal to the number
of genes.}

\item{singleGroup}{Character string specifying the name of a single group
within groups to plot.  Individual blot plots are generated for each gene
on a single set of axes.}

\item{singleColor}{Character string specifying the color to use for the
singleGroup blot plot.}

\item{order}{Boolean specifying whether the groups should be arranged in
alphabetical factor order prior to plotting.  If FALSE, the current order
of levels is used.}

\item{save}{Boolean specifying whether to save the resultant plot as a .tiff
file.}
}
\value{
Blot plots for the specified groups and genes in a new window.
}
\description{
Takes ExpressionSet object and generates blot plots for the specified groups
and genes.
}

