% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorData.R
\name{cData}
\alias{cData}
\title{Display information in the colorData slot}
\usage{
cData(cellData)
}
\description{
This function returns the information stored in the colorData slot of a Hocus object.
}

