% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusterCells.R
\name{clusterCells}
\alias{clusterCells}
\title{Perform clustering analysis on the samples in the expression matrix.}
\usage{
clusterCells(cellData, k = 3, methods = c("hierarchical", "kmeans", "pam"),
  hier_dist = "euclidean", hier_clust = "ward")
}
\arguments{
\item{cellData}{ExpressionSet object created with readCells (and preferably
transformed with prepCells).  It is also helpful to first run
reduceGenes_var and reduceGenes_pca.}

\item{k}{The number of desired clusters into which the samples are to be
grouped.  The function calcGap can be used to obtain an unbiased estimate
of the number of clusters.}

\item{methods}{Vector of character strings specifying which of the three
clustering methods to perform.  All three can be specified, or a subset of
the three.}

\item{hier_dist}{Character string specifying how to compute the distance
matrix for 'hierarchical.'  Equivalent to the 'method' parameter within the
dist function.}

\item{hier_clust}{Character string specifying the clustering method for
'hierarchical.'  Equivalent to the 'method' parameter within the hclust
function.}
}
\value{
For each clustering method specified, a column is added to pData
  containing the cluster information for each sample (clusters are designated
  by roman numerals).
}
\description{
Takes ExpressionSet object and clusters the samples using hierarchical,
k-means, or pam clustering.
}

