% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colorData.R
\name{colorData}
\alias{colorData}
\title{Display information in the colorData slot}
\usage{
colorData(cellData)
}
\description{
This function returns the information stored in the colorData slot of a Hocus object.
}

