% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gapStats.R
\name{gapStats}
\alias{gapStats}
\title{Unbiased estimate of the number of cell or gene clusters using the gap
statistic.}
\usage{
gapStats(cellData, gene_clust = FALSE, fun = "kmeans", max_clust = 25,
  boot = 100, plot = TRUE, save = FALSE, print = TRUE)
}
\arguments{
\item{cellData}{ExpressionSet object created with readCells (and preferably
transformed with prepCells).  It is also helpful to first run
reduceGenes_var and reduceGenes_pca.}

\item{gene_clust}{Boolean specifying whether the gap statistic should be
calculated for the samples or genes.  TRUE calculates for the cells, FALSE
for the genes.}

\item{fun}{Character string specifying whether the gap statistic should be
calculated for kmeans, pam, or hierarchical clustering.  Possible values
are kmeans, pam, or hclust. clustering methods to perform.  All three can
be specified, or a subset of the three.}

\item{max_clust}{Integer specifying the maximum possible number of clusters
in the dataset.  Set higher than the expected value. matrix for
'hierarchical.'  Equivalent to the 'method' parameter within the dist
function.}

\item{boot}{Integer specifying the number of bootstrap iterations to perform
when calculating the gap statistic. 'hierarchical.'  Equivalent to the
'method' parameter within the hclust function.}

\item{plot}{Boolean specifying whether a plot of the gap values vs the number
of clusters should be produced.}

\item{save}{Boolean specifying whether the plot should be saved.}

\item{print}{Boolean specifying whether the optimal number of clusters should
be printed in the terminal window.}
}
\value{
The optimal number of clusters calculated from the gap statistic with
  the given parameters.  A new column is added to pData indicating the
  optimal number of cell or gene clusters for the chosen clustering method.
}
\description{
Takes ExpressionSet object and calculates the optimal number of kmeans, pam,
or hierarchical clusters for the samples or genes using the gap statistic.
}

