% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logData.R
\name{logData}
\alias{logData}
\title{Display information in the logData slot}
\usage{
logData(cellData)
}
\description{
This function returns the information stored in the logData slot of a Hocus object.
}

