% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca1D.R
\name{pca1D}
\alias{pca1D}
\title{}
\usage{
pca1D(cellData, center = TRUE, scale = FALSE, PC = 1, ICA = FALSE,
  groups, values, gene, colors, logNumeric = TRUE, alpha = 0.7,
  dotsize = 5, bubble = FALSE, bubbleSizes, save = FALSE)
}
\description{

}

