% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepCells.R
\name{prepCells}
\alias{prepCells}
\title{Prepare matrix of expression values for downstream analyses}
\usage{
prepCells(cellData, LOD = 1, norm = TRUE, batchGroup)
}
\arguments{
\item{cellData}{ExpressionSet object created with readCells.}

\item{LOD}{Limit of detection (numeric).  All expression values below this
limit will be set to zero prior to the log2 transformation.  LOD is best
determined by spiking in a dilution series of RNA standards with known
concentration.}

\item{norm}{Boolean specifying whether to normalize the expression values of
each gene across all samples in the ExpressionSet object using the median
of the geometric means.  This helps with making expression values more
comparable across samples. See explanation of the
estimateSizeFactorsForMatrix function within the DESeq package
documentation for more explanation.}

\item{batchGroup}{Character string specifying the name of the column header
in pData where the batch group information is stored.  All samples within a
particular batch group should have the same designation.  If no character
string is specified, batch is not run.  E.g. if there are four samples and
two batch groups, the information in the batch group column could look like
c('Batch1','Batch1','Batch2','Batch2').}
}
\value{
ExpressionSet object modified according to the optional parameters
  specified above.  Note that it is highly recommended that you at least
  log-transform the data prior to any further downstream analysis.
}
\description{
Takes ExpressionSet object and (optionally) 1) normalizes the expression
levels across samples using the median of the geometric means method found
within the DESeq package, 2) transforms the values in the matrix of
expression values to a log2 scale, 3) sets expression values below the set
limit of detection (LOD) to zero prior to the log2 transformation, and 4)
provides an option to eliminate potential batch effects specified in a
phenoData column using the ComBat method within the sva package.
}

