% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readCells.R
\name{readCells}
\alias{readCells}
\title{Load all experiment data into R}
\usage{
readCells(assay, pheno, feature, experiment, color)
}
\arguments{
\item{assay}{Character string specifying the path and filename of a
tab-delimited .txt file containing a matrix of expression values (e.g.
counts, FPKMs, TPMs, etc.) with cells in columns and genes in rows.  Be
sure to include cell and gene names as header and row names!}

\item{pheno}{Character string specifying the path and filename tab-delimited
.txt file containing information about the cells in separate columns.  Be
sure that the row names in this file match the column names in assayData
(can be a subset)}

\item{feature}{Character string specifying the path and filename of a
tab-delimited .txt file containing information about the genes in separate
columns.  Be sure that the row names in this file match the row names in
assayData (can be a subset)}

\item{experiment}{Character string specifying the path and filename of a
tab-delimited .txt file containing additional information about all the
column headers in the phenoData file in separate columns.  Be sure that the
row names in this file match the column headers in the phenoData file.}
}
\value{
ExpressionSet object containg matrix of expression values, phenoData,
  assayData, and experimentData.
}
\description{
Takes separate .txt files containing assay, pheno, feature, and experiment
data and combines them into a single ExpressionSet object (uses Bioconductor
ExpressionSet class format).
}

