% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reduceGenes_var.R
\name{reduceGenes_var}
\alias{reduceGenes_var}
\title{Reduce number of genes in expression matrix based on variance}
\usage{
reduceGenes_var(cellData, exprGenes = TRUE, exprThresh = 1,
  cellThresh = 2, varThresh = TRUE, cv = 0.5, seuratThresh = FALSE,
  z_cutoff = 1.5, ...)
}
\arguments{
\item{cellData}{ExpressionSet object created with readCells (and preferably
transformed with prepCells)}

\item{exprGenes}{Boolean specifying whether to remove genes that are
expressed above the expression threshold (exprThresh) in fewer than the
number of samples specified with cellThresh.}

\item{exprThresh}{Numeric specifying the the threshold for considering a gene
to be expressed.  Note that if prepCells has been run, expression values
will be on a log2 scale, so exprThresh = 1 corresponds to an expression
level of 2 in untransformed space.}

\item{cellThresh}{Integer indicating the number of cells in which a gene has
to be expressed above the expression threshold (exprThresh) in order to be
kept.}

\item{varThresh}{Boolean specifying whether to remove genes whose CV across
samples is less than the threshold specified with cv.  If exprGenes and
varThresh are both TRUE, exprGenes is performed prior to varThresh.}

\item{cv}{Numeric specifying the CV threshold for varThresh.  Values ranging
from 0.1 to 0.5 are typically effective.}

\item{seuratThresh}{Boolean specifying whether the Seurat package's method of
elminating genes by variance is to be used.  See the Seurat package
documentation for more details.  If exprThresh, varThresh, and seuratThresh
are all TRUE, they are performed in that order.}

\item{z_cutoff}{Numeric specifying the z-score cutoff for the Seurat method.}

\item{...}{Pass more options to the Seurat function mean.var.plot()}
}
\value{
ExpressionSet object with genes removed from the expression matrix
  according to the optional parameters specified above.  Note that the
  original list of genes will still be present within fData.  Genes that pass
  filter will be stored in fData as TRUE, genes that do not pass filter will
  be stored as FALSE.
}
\description{
Takes ExpressionSet object and (optionally) 1) eliminates genes that are
expressed below the LOD in a specified number of samples, 2) eliminates genes
whose coefficient of variation (CV) across cells is below a set limit, and 3)
eliminates genes using the CV z-score/binning method within the Seurat
package.
}

