% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setColors.R
\name{setColors}
\alias{setColors}
\title{Add colors to colorData slot}
\usage{
setColors(cellData, groups, colors, auto = FALSE, palette = "rainbow")
}
\description{
Add colors for specified groups to the colorData slot in a Hocus object.
}

