% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subGroups.R
\name{subGroups}
\alias{subGroups}
\title{Select a subset of groups in phenoData}
\usage{
subGroups(cellData, groups, names)
}
\description{
Select a subset of specified groups in the phenoData slot of a Hocus object while retaining associated color info in colorData.
}

